feat(MarcaController): implementado controle de fluxo nos endpoint

+++ app/Http/Controllers/MarcaController.php
@@ -59,6 +59,9 @@ class MarcaController extends Controller
     {
         //Sugestão de tipo, type hinting do php
         $marca = $this->marca->find($id);
+        if($marca === null){
+            return ['erro' => 'Recurso pesquisado não existe'];
+        };
         return $marca;
     }
 
@@ -89,6 +92,11 @@ class MarcaController extends Controller
        */
 
        $marca = $this->marca->find($id);
+
+       if($marca === null){
+            return ['erro' => 'Impossível realizar a atualização. O recurso solicitado não existe'];
+       };
+
        $marca->update($request->all());
        return $marca;
     }
@@ -102,6 +110,11 @@ class MarcaController extends Controller
     public function destroy($id)
     {
         $marca = $this->marca->find($id);
+
+        if($marca === null){
+            return ['erro' => 'Impossível realizar a exclusão. O recurso solicitado não existe'];
+        };
+
         $marca->delete();
         return ['msg' => 'A marca foi removida com sucesso!'];
     }
